@@ -1,72 +1,69 @@
 {
     "name": "iihf-web",
-    "version": "1.2.2",
+    "version": "1.3.0",
     "scripts": {
         "ng": "ng",
         "start": "ng serve --proxy-config proxy.conf.json",
-        "startAOT": "ng serve --aot=true --proxy-config proxy.conf.json",
         "build": "ng build --configuration=devou",
-        "buildProd": "ng build --configuration=production",
-        "buildDockerTest": "ng build --configuration=docker",
-        "buildDockerDemo": "ng build --configuration=demo",
-        "zipDist": "zip -r dist/iihf-web.zip dist/iihf-web",
-        "buildAndZipDist": "npm run build && npm run zipDist",
+        "buildTest": "ng build --configuration=test && npm run zipDist",
+        "buildDemo": "ng build --configuration=demo && npm run zipDist",
+        "buildProd": "ng build --configuration=production && npm run zipDist",
         "test": "ng test",
         "lint": "ng lint",
-        "e2e": "ng e2e"
+        "e2e": "ng e2e",
+        "zipDist": "cd dist && zip -r ../distzip/iihf-eps-web.zip *"
     },
     "private": true,
     "dependencies": {
-        "@angular/animations": "^7.2.8",
-        "@angular/cdk": "^7.3.3",
-        "@angular/common": "~7.2.0",
-        "@angular/compiler": "~7.2.0",
-        "@angular/core": "~7.2.0",
+        "@angular/animations": "^8.2.10",
+        "@angular/cdk": "^8.2.3",
+        "@angular/common": "~8.2.10",
+        "@angular/compiler": "~8.2.10",
+        "@angular/core": "~8.2.10",
         "@angular/flex-layout": "^7.0.0-beta.23",
-        "@angular/forms": "~7.2.0",
-        "@angular/material": "^7.3.3",
-        "@angular/material-moment-adapter": "^7.3.3",
-        "@angular/platform-browser": "~7.2.0",
-        "@angular/platform-browser-dynamic": "~7.2.0",
-        "@angular/router": "~7.2.0",
+        "@angular/forms": "~8.2.10",
+        "@angular/material": "^8.2.3",
+        "@angular/material-moment-adapter": "^8.2.3",
+        "@angular/platform-browser": "~8.2.10",
+        "@angular/platform-browser-dynamic": "~8.2.10",
+        "@angular/router": "~8.2.10",
         "@ngx-translate/core": "^11.0.1",
         "@ngx-translate/http-loader": "^4.0.0",
-        "chalk": "^2.4.2",
         "classlist.js": "^1.1.20150312",
-        "core-js": "^2.5.4",
+        "core-js": "^3.2.1",
         "hammerjs": "^2.0.8",
         "moment": "^2.24.0",
         "ng2-pdf-viewer": "^5.3.4",
-        "rxjs": "^6.3.3",
-        "screenfull": "^4.0.1",
-        "tslib": "^1.9.0",
-        "tslint-clean-code": "^0.2.9",
-        "tslint-consistent-codestyle": "^1.15.1",
+        "rxjs": "^6.5.3",
+        "tslib": "^1.10.0",
+        "tslint-clean-code": "^0.2.10",
+        "tslint-consistent-codestyle": "^1.16.0",
         "tslint-sonarts": "^1.9.0",
-        "web-animations-js": "^2.3.1",
-        "zone.js": "~0.8.26"
+        "tslint-strict-null-checks": "^1.0.1",
+        "web-animations-js": "^2.3.2",
+        "zone.js": "~0.10.2"
     },
     "devDependencies": {
-        "@angular-devkit/build-angular": "^0.13.9",
-        "@angular/cli": "~7.3.4",
-        "@angular/compiler-cli": "~7.2.0",
-        "@angular/language-service": "~7.2.0",
-        "@biesbjerg/ngx-translate-extract": "^2.3.4",
-        "@types/jasmine": "~2.8.8",
-        "@types/jasminewd2": "~2.0.3",
-        "@types/node": "~8.9.4",
+        "@angular-devkit/build-angular": "^0.803.8",
+        "@angular/cli": "~8.3.8",
+        "@angular/compiler-cli": "~8.2.10",
+        "@angular/language-service": "~8.2.10",
+        "@biesbjerg/ngx-translate-extract": "^4.2.0",
+        "@types/jasmine": "~3.4.3",
+        "@types/jasminewd2": "~2.0.8",
+        "@types/node": "~12.7.12",
         "chai": "^4.2.0",
-        "codelyzer": "^4.5.0",
-        "jasmine-core": "~2.99.1",
+        "codelyzer": "^5.1.2",
+        "jasmine-core": "~3.5.0",
         "jasmine-spec-reporter": "~4.2.1",
-        "karma": "~4.0.0",
-        "karma-chrome-launcher": "~2.2.0",
-        "karma-coverage-istanbul-reporter": "~2.0.1",
-        "karma-jasmine": "~1.1.2",
-        "karma-jasmine-html-reporter": "^0.2.2",
-        "protractor": "~5.4.0",
-        "ts-node": "~7.0.0",
-        "tslint": "~5.11.0",
-        "typescript": "~3.2.2"
+        "karma": "~4.3.0",
+        "karma-chrome-launcher": "~3.1.0",
+        "karma-coverage-istanbul-reporter": "~2.1.0",
+        "karma-jasmine": "~2.0.1",
+        "karma-jasmine-html-reporter": "^1.4.2",
+        "protractor": "~5.4.2",
+        "ts-node": "~8.4.1",
+        "tslint": "~5.20.0",
+        "typescript": "^3.5.3"
     }
 }
