@@ -1,10 +1,10 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { ReactiveFormsModule } from '@angular/forms';
 import { NoopAnimationsModule } from '@angular/platform-browser/animations';
-import { AbstractInputsModule } from '../abstract-inputs.module';
 
 import { TRANSLATE_WRAPPER_TOKEN } from '../../hazelnut-common/interfaces';
 import { NoopTranslationsService } from '../../hazelnut-common/services/testing/noop-translation.service';
+import { AbstractInputsModule } from '../abstract-inputs.module';
 import { InputNumberRangeComponent } from './input-number-range.component';
 
 describe('InputNumberRangeComponent', () => {
@@ -13,15 +13,19 @@ describe('InputNumberRangeComponent', () => {
 
     beforeEach(async(() => {
         TestBed.configureTestingModule({
-            imports: [
-                AbstractInputsModule,
-                NoopAnimationsModule,
-                ReactiveFormsModule,
-            ],
-            providers: [
-                {provide: TRANSLATE_WRAPPER_TOKEN, useClass: NoopTranslationsService},
-            ]
-        }).compileComponents();
+                   imports: [
+                       AbstractInputsModule,
+                       NoopAnimationsModule,
+                       ReactiveFormsModule,
+                   ],
+                   providers: [
+                       {
+                           provide: TRANSLATE_WRAPPER_TOKEN,
+                           useClass: NoopTranslationsService
+                       },
+                   ]
+               })
+               .compileComponents();
     }));
 
     beforeEach(() => {
@@ -31,6 +35,7 @@ describe('InputNumberRangeComponent', () => {
     });
 
     it('should create', () => {
-        expect(component).toBeTruthy();
+        expect(component)
+            .toBeTruthy();
     });
 });
