@@ -6,6 +6,7 @@ import { UserDataService } from '../../../../services/data/user-data.service';
 import { NotificationService } from '../../../../services/notification.service';
 import { fadeEnterLeave } from '../../../hazelnut-common/animations';
 import { StringUtils } from '../../../hazelnut-common/hazelnut';
+import { Filter } from '../../../hazelnut-common/models';
 import { CoreTableService } from '../../core-table.service';
 import { TableColumn } from '../../models/table-column.model';
 import { TableFilterType } from '../../models/table-filter-type.enum';
@@ -17,28 +18,9 @@ import { TableFilterType } from '../../models/table-filter-type.enum';
     animations: [fadeEnterLeave]
 })
 export class CoreTableFilterComponent implements OnInit {
-    @Input() public columnConfig: TableColumn = new TableColumn({columnDef: '', label: ''});
-    @Output() public resetFilters = new EventEmitter<boolean>();
-    private _filtersElement: FormGroup = null;
-    public showSearchIcon = true;
-    public filterTypeEnum: typeof TableFilterType = TableFilterType;
-    public userList$ = this.userDataService.getUsers();
-    public categoryList$ = [];
-
-    public constructor(private readonly coreTableService: CoreTableService,
-                       private readonly userDataService: UserDataService,
-                       private readonly notificationService: NotificationService,
-                       private readonly businessAreaService: BusinessAreaService,
-    ) {
-
-    }
 
     public get filterDelay(): number {
         return this.coreTableService && this.coreTableService.configuration ? this.coreTableService.configuration.filterDebounceTime : 0;
-    }
-
-    public get filtersElement(): FormGroup {
-        return this._filtersElement;
     }
 
     public get control(): AbstractControl {
@@ -49,14 +31,33 @@ export class CoreTableFilterComponent implements OnInit {
         return this.columnConfig.filter.width ? `w-${this.columnConfig.filter.width}` : 'w-100';
     }
 
+    public get filtersElement(): FormGroup {
+        return this._filtersElement;
+    }
+
+    @Input() public columnConfig: TableColumn = new TableColumn({
+        columnDef: '',
+        label: ''
+    });
+    @Output() public resetFilters = new EventEmitter<boolean>();
+    public showSearchIcon = true;
+    public filterTypeEnum: typeof TableFilterType = TableFilterType;
+    public userList$ = this.userDataService.getUsers();
+    public categoryList$ = [];
+
+    private _filtersElement: FormGroup = null;
+
+    public constructor(private readonly coreTableService: CoreTableService,
+                       private readonly userDataService: UserDataService,
+                       private readonly notificationService: NotificationService,
+                       private readonly businessAreaService: BusinessAreaService, ) {
+
+    }
+
     public ngOnInit(): void {
         const elementName = this.columnConfig.filterElement;
-        const formGroupOptions: { [id: string]: FormControl } = {};
-        formGroupOptions[elementName] = new FormControl(
-            this.columnConfig.filter && this.columnConfig.filter.predefinedValue ?
-                this.columnConfig.filter.predefinedValue[0] :
-                null,
-        );
+        const formGroupOptions: {[id: string]: FormControl} = {};
+        formGroupOptions[elementName] = new FormControl(this.columnConfig.filter && this.columnConfig.filter.predefinedValue ? this.columnConfig.filter.predefinedValue[0] : null, );
         this._filtersElement = new FormGroup(formGroupOptions);
         this.processFormChanges();
 
@@ -67,57 +68,53 @@ export class CoreTableFilterComponent implements OnInit {
         this.setPredefinedFilterValuesIntoTableFilters();
     }
 
-    private loadCategoryList() {
+    public clearFilters(): void {
+        this.resetFilters.next(true);
+        this.coreTableService.clearFilters();
+    }
+
+    private loadCategoryList(): void {
         this.businessAreaService.listCategories()
             .subscribe((data) => {
                 this.categoryList$ = data.content;
             });
     }
 
-    public clearFilters() {
-        this.resetFilters.next(true);
-        this.coreTableService.clearFilters();
-    }
-
     private processFormChanges(): void {
         const elementName = this.columnConfig.filterElement;
-        this.filtersElement.valueChanges.pipe(debounceTime(this.filterDelay)).subscribe((values: any) => {
-            if (this.filtersElement.status === 'VALID') {
-                const newValue = values[elementName];
-                this.coreTableService.addFilter(this.columnConfig, newValue);
-            }
-        });
+        this.filtersElement.valueChanges.pipe(debounceTime(this.filterDelay))
+            .subscribe((values: any) => {
+                if (this.filtersElement.status === 'VALID') {
+                    const newValue = values[elementName];
+                    this.coreTableService.addFilter(this.columnConfig, newValue);
+                }
+            });
     }
 
-    private setPredefinedFilterValuesIntoTableFilters() {
+    private setPredefinedFilterValuesIntoTableFilters(): void {
         const configuration = this.coreTableService.configuration;
-        const isFilterFromPredefinedFilters = configuration && configuration.predefinedFilters
-            && configuration.predefinedFilters.find((filter) =>
-                filter.property === StringUtils.convertCamelToSnakeUpper(this.columnConfig.columnDef)
-            );
+        const isFilterFromPredefinedFilters = configuration &&
+            configuration.predefinedFilters &&
+            configuration.predefinedFilters.find((filter: Filter) => filter.property === StringUtils.convertCamelToSnakeUpper(this.columnConfig.columnDef));
         const isPredefinedNumberValue = this.columnConfig.filter.type === 'number';
-        const isPredefinedDateValue = this.columnConfig.config
-            && this.columnConfig.config.filter
-            && this.columnConfig.config.filter.valueType === 'DATE_TIME';
+        const isPredefinedDateValue = this.columnConfig.config && this.columnConfig.config.filter && this.columnConfig.config.filter.valueType === 'DATE_TIME';
         const isPredefinedTrafficLightValue = this.columnConfig.columnDef === 'trafficLight';
         if (isFilterFromPredefinedFilters) {
             if (isPredefinedNumberValue) {
                 this.setFilterElementByPredefinedNumberValues();
             } else if (isPredefinedDateValue) {
-               // TODO apply set date filter
+                // TODO apply set date filter
             } else if (isPredefinedTrafficLightValue) {
                 this.setFilterElementByPredefinedTrafficLightValues();
             } else {
                 this.filtersElement.controls[this.columnConfig.filterElement].patchValue(configuration.predefinedFilters
-                    .find((filter) =>
-                        filter.property === StringUtils.convertCamelToSnakeUpper(this.columnConfig.columnDef))
-                    .value
-                );
+                                                                                                      .find((filter: Filter) => filter.property ===
+                                                                                                          StringUtils.convertCamelToSnakeUpper(this.columnConfig.columnDef)).value);
             }
         }
     }
 
-    private setFilterElementByPredefinedDateValues() {
+    private setFilterElementByPredefinedDateValues(): void {
         // TODO fix ExpressionChangedAfterItHasBeenCheckedError
         const lowerFilter = this.getNumberFilterFromPredefinedFiltersByOperator('GOE');
         const higherFilter = this.getNumberFilterFromPredefinedFiltersByOperator('LOE');
@@ -127,7 +124,7 @@ export class CoreTableFilterComponent implements OnInit {
         });
     }
 
-    private setFilterElementByPredefinedNumberValues() {
+    private setFilterElementByPredefinedNumberValues(): void {
         const lowerFilter = this.getNumberFilterFromPredefinedFiltersByOperator('GOE');
         const higherFilter = this.getNumberFilterFromPredefinedFiltersByOperator('LOE');
         this.filtersElement.controls[this.columnConfig.filterElement].patchValue({
@@ -136,27 +133,29 @@ export class CoreTableFilterComponent implements OnInit {
         });
     }
 
-    private setFilterElementByPredefinedTrafficLightValues() {
-        this.filtersElement.controls[this.columnConfig.filterElement].patchValue(
-            ['RED', 'GREEN', 'AMBER', 'NONE']
-                .map((color) => this.getTrafficLightFromPredefinedFiltersByColor(color))
-                .filter((filter) => filter !== undefined)
-                .map((filter) => filter.value.toString().toLowerCase())
-        );
+    private setFilterElementByPredefinedTrafficLightValues(): void {
+        this.filtersElement.controls[this.columnConfig.filterElement].patchValue([
+            'RED',
+            'GREEN',
+            'AMBER',
+            'NONE'
+        ]
+            .map((color: string) => this.getTrafficLightFromPredefinedFiltersByColor(color))
+            .filter((filter: Filter) => filter !== undefined)
+            .map((filter: Filter) => filter.value.toString()
+                                           .toLowerCase()));
     }
 
-    private getNumberFilterFromPredefinedFiltersByOperator(operator: string) {
+    private getNumberFilterFromPredefinedFiltersByOperator(operator: string): Filter {
         return this.coreTableService.configuration.predefinedFilters
-            .filter((filter) =>
-                filter.property === StringUtils.convertCamelToSnakeUpper(this.columnConfig.columnDef))
-            .find((filter) => filter.operator === operator);
+                   .filter((filter: Filter) => filter.property === StringUtils.convertCamelToSnakeUpper(this.columnConfig.columnDef))
+                   .find((filter: Filter) => filter.operator === operator);
     }
 
-    private getTrafficLightFromPredefinedFiltersByColor(colorValue: string) {
+    private getTrafficLightFromPredefinedFiltersByColor(colorValue: string): Filter {
         return this.coreTableService.configuration.predefinedFilters
-            .filter((filter) =>
-                filter.property === StringUtils.convertCamelToSnakeUpper(this.columnConfig.columnDef))
-            .find((filter) => filter.value === colorValue);
+                   .filter((filter: Filter) => filter.property === StringUtils.convertCamelToSnakeUpper(this.columnConfig.columnDef))
+                   .find((filter: Filter) => filter.value === colorValue);
     }
 
 }
