@@ -9,18 +9,14 @@ import { TableCellCustomized } from '../../interfaces/table-cell-customized';
 export class TableCellClickableComponent implements TableCellCustomized, OnInit {
     @Input() public row: any;
     @Input() public params: {
-        contentPath: string[];
-        content: string;
-        link?: string;
-        linkParts?: string[];
-        callback?: (row: any) => void;
+        contentPath: string[]; content: string; link?: string; linkParts?: string[]; callback?(row: any): void;
     };
 
     public title: string;
     public component: any;
 
     public ngOnInit(): void {
-        // empty
+        // Empty
     }
 
     public onClick(): void {
