@@ -5,22 +5,22 @@ import { Directive, EventEmitter, HostListener, Output } from '@angular/core';
 })
 export class DragDropDirective {
 
-    @Output() public onFileDropped = new EventEmitter<any>();
+    @Output() public readonly onFileDropped = new EventEmitter<any>();
 
     @HostListener('dragover', ['$event'])
-    public onDragOver(event) {
+    public onDragOver(event): void {
         event.preventDefault();
         event.stopPropagation();
     }
 
     @HostListener('dragleave', ['$event'])
-    public onDragLeave(event) {
+    public onDragLeave(event): void {
         event.preventDefault();
         event.stopPropagation();
     }
 
     @HostListener('drop', ['$event'])
-    public ondrop(event) {
+    public ondrop(event): void {
         event.preventDefault();
         event.stopPropagation();
         const files = event.dataTransfer.files;
