@@ -10,11 +10,23 @@ describe('StringUtils', () => {
     describe('Join', () => {
         it('It should join string by delimiter and append prefix and postfix', () => {
             const notArray: any[] | any = 'Jean-Claude Van Damme';
-            const testCase = ['hello', 'my', 'world'];
-            expect(StringUtils.join(notArray, '///', '<<', '>>')).to.equal(`<<${notArray}>>`);
-            expect(StringUtils.join(testCase, ' ', '<<', '>>')).to.equal('<<hello my world>>');
-            expect(StringUtils.join([], '-', '<<', '>>')).to.equal('<<>>');
-            expect(StringUtils.join([], '/')).to.equal('');
+            const testCase = [
+                'hello',
+                'my',
+                'world'
+            ];
+            expect(StringUtils.join(notArray, '///', '<<', '>>'))
+                .to
+                .equal(`<<${notArray}>>`);
+            expect(StringUtils.join(testCase, ' ', '<<', '>>'))
+                .to
+                .equal('<<hello my world>>');
+            expect(StringUtils.join([], '-', '<<', '>>'))
+                .to
+                .equal('<<>>');
+            expect(StringUtils.join([], '/'))
+                .to
+                .equal('');
         });
     });
     describe('RemoveAccented', () => {
@@ -22,21 +34,50 @@ describe('StringUtils', () => {
         const finalTestString = testString + testString.toUpperCase();
         const finalResultString = resultString + resultString.toUpperCase();
         it('It should remove accented characters from string', () => {
-            expect(StringUtils.removeAccentedCharacters(finalTestString)).to.be.equal(finalResultString);
-            expect(StringUtils.removeAccentedCharacters(finalResultString)).to.be.equal(finalResultString);
-            expect(StringUtils.removeAccentedCharacters(MockData.characters)).to.be.equal(MockData.characters);
-            expect(StringUtils.removeAccentedCharacters(notString)).to.equal(notString);
+            expect(StringUtils.removeAccentedCharacters(finalTestString))
+                .to
+                .be
+                .equal(finalResultString);
+            expect(StringUtils.removeAccentedCharacters(finalResultString))
+                .to
+                .be
+                .equal(finalResultString);
+            expect(StringUtils.removeAccentedCharacters(MockData.characters))
+                .to
+                .be
+                .equal(MockData.characters);
+            expect(StringUtils.removeAccentedCharacters(notString))
+                .to
+                .equal(notString);
         });
     });
 
     describe('GetLastPart', () => {
         it('It should return last part of string divided by delimiter', () => {
-            expect(StringUtils.getLastPart('ababababa', 'b')).to.be.equal('a');
-            expect(StringUtils.getLastPart('ababababaa', 'b')).to.be.equal('aa');
-            expect(StringUtils.getLastPart('i am here')).to.be.equal('here');
-            expect(StringUtils.getLastPart('i ambhere', 'b')).to.be.equal('here');
-            expect(StringUtils.getLastPart('', '')).to.be.equal('');
-            expect(StringUtils.getLastPart('')).to.be.equal('');
+            expect(StringUtils.getLastPart('ababababa', 'b'))
+                .to
+                .be
+                .equal('a');
+            expect(StringUtils.getLastPart('ababababaa', 'b'))
+                .to
+                .be
+                .equal('aa');
+            expect(StringUtils.getLastPart('i am here'))
+                .to
+                .be
+                .equal('here');
+            expect(StringUtils.getLastPart('i ambhere', 'b'))
+                .to
+                .be
+                .equal('here');
+            expect(StringUtils.getLastPart('', ''))
+                .to
+                .be
+                .equal('');
+            expect(StringUtils.getLastPart(''))
+                .to
+                .be
+                .equal('');
 
         });
     });
