@@ -1,15 +1,14 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
-import { SafePipe } from './safe.pipe';
 import { ThousandDelimiterPipe } from './thousand-delimiter.pipe';
 import { VenuePipe } from './venue.pipe';
 
 @NgModule({
-    declarations: [VenuePipe, ThousandDelimiterPipe, SafePipe],
+    declarations: [VenuePipe, ThousandDelimiterPipe],
     imports: [
         CommonModule,
     ],
-    exports: [VenuePipe, ThousandDelimiterPipe, SafePipe]
+    exports: [VenuePipe, ThousandDelimiterPipe]
 })
 export class PipesModule {
 }
