@@ -2,7 +2,7 @@ import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
 import { catchError, map } from 'rxjs/operators';
-import { HazelnutConfig } from '../../hazlenut/hazelnut-common/config/hazelnut-config';
+import { hazelnutConfig } from '../../hazlenut/hazelnut-common/config/hazelnut-config';
 import { NotificationService } from '../notification.service';
 import { ProjectService } from '../project.service';
 import { ProjectUserService } from '../storage/project-user.service';
@@ -14,7 +14,7 @@ import { ProjectUserService } from '../storage/project-user.service';
 // TODO: remove project Service or add function to porject service
 export class ImagesService extends ProjectService<any> {
 
-    public constructor(http: HttpClient, notificationService: NotificationService, userService: ProjectUserService, ) {
+    public constructor(http: HttpClient, notificationService: NotificationService, userService: ProjectUserService,) {
         super(http, 'images', notificationService, userService);
     }
 
@@ -28,14 +28,14 @@ export class ImagesService extends ProjectService<any> {
         headers = headers.set('token', this.userService.instant.authToken);
         return this.post({
             headers,
-            url: `${HazelnutConfig.URL_API}/internal/images`,
+            url: `${hazelnutConfig.URL_API}/internal/images`,
             mapFunction: (e) => e,
             body: formData,
         });
     }
 
     public getImage(imageName: string): Observable<Blob> {
-        return this.http.get(`${HazelnutConfig.URL_API}/internal/images/0/${imageName}`, {
+        return this.http.get(`${hazelnutConfig.URL_API}/internal/images/0/${imageName}`, {
             headers: this.getHeader(),
             responseType: 'blob',
         }).pipe(
