@@ -1,6 +1,6 @@
 import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { Observable } from 'rxjs';
-import { HazelnutConfig } from '../hazlenut/hazelnut-common/config/hazelnut-config';
+import { hazelnutConfig } from '../hazlenut/hazelnut-common/config/hazelnut-config';
 import { AbstractService } from '../hazlenut/hazelnut-common/services';
 import { NotificationService } from './notification.service';
 import { ProjectUserService } from './storage/project-user.service';
@@ -17,14 +17,6 @@ export class ProjectService<T> extends AbstractService<T> {
         super(http, notificationService, url);
     }
 
-    protected getHeader(): HttpHeaders {
-        let headers = new HttpHeaders();
-        headers = headers.set('device-id', this.userService.instant.deviceId);
-        headers = headers.set('token', localStorage.getItem(this.userService.login) ? JSON.parse(localStorage.getItem(this.userService.login)).authToken : '');
-        headers = headers.set('Content-Type', 'application/json');
-        return headers;
-    }
-
     /**
      * Function returns list of result from browse API
      * @param id - id of searched object
@@ -33,9 +25,17 @@ export class ProjectService<T> extends AbstractService<T> {
     public getFactItemDetail(id: number | string, projectId: number | string): Observable<T> {
         const realIds = id && projectId ? `/${id}/${projectId}` : '';
         return this.get({
-            url: `${HazelnutConfig.URL_API}/${this.urlKey}${realIds}`,
+            url: `${hazelnutConfig.URL_API}/${this.urlKey}${realIds}`,
             mapFunction: this.extractDetail
         });
+    }
+
+    protected getHeader(): HttpHeaders {
+        let headers = new HttpHeaders();
+        headers = headers.set('device-id', this.userService.instant.deviceId);
+        headers = headers.set('token', localStorage.getItem(this.userService.login) ? JSON.parse(localStorage.getItem(this.userService.login)).authToken : '');
+        headers = headers.set('Content-Type', 'application/json');
+        return headers;
     }
 
 }
